refactor: 게시글의 id 관련 참조들 수정

+++ src/main/java/codesquad/codestagram/repository/ArticleRepository.java
@@ -9,7 +9,7 @@ public interface ArticleRepository {
 
     Article save(Article article);
 
-    Optional<Article> findByArticleId(Long id);
+    Optional<Article> findByArticleId(int id);
 
     List<Article> findAllArticles();
 }
